lecture_5/examples/django-rest/src/requirements.txt
coreapi
django
djangorestframework
factory_boy
ipython
pygments  # We'll be using this for the code highlighting
pycodestyle
pylint
pylint-django
